+++ src/main/java/chess/dto/board/BoardDTO.java
@@ -1,4 +1,4 @@
-package chess.dto;
+package chess.dto.board;
 
 import chess.domain.board.*;
 import chess.domain.piece.TeamType;
@@ -18,21 +18,21 @@ public class BoardDTO {
         this.isCheckmate = isCheckmate;
     }
 
-    public static BoardDTO from(ChessBoard chessBoard, TeamType currentTeamType) {
+    public static BoardDTO of(ChessBoard chessBoard, TeamType currentTeamType) {
         Map<Coordinate, Cell> cells = chessBoard.getCells();
         List<RowDTO> rows = new ArrayList<>();
         List<Rank> ranks = Arrays.stream(Rank.values())
                 .sorted(Comparator.reverseOrder())
                 .collect(Collectors.toList());
         for (Rank rank : ranks) {
-            List<PieceDTO> piecesByRow = writePiecesByRow(rank, cells);
+            List<PieceDTO> piecesByRow = aggregatePiecesByRow(rank, cells);
             RowDTO rowDTO = new RowDTO(piecesByRow);
             rows.add(rowDTO);
         }
         return new BoardDTO(rows, currentTeamType, chessBoard.isKingCheckmate());
     }
 
-    private static List<PieceDTO> writePiecesByRow(Rank rank, Map<Coordinate, Cell> cells) {
+    private static List<PieceDTO> aggregatePiecesByRow(Rank rank, Map<Coordinate, Cell> cells) {
         List<PieceDTO> piecesByRow = new ArrayList<>();
         for (File file : File.values()) {
             Coordinate coordinate = new Coordinate(file, rank);
